requirements.txt
Flask
Flask-RESTful
python-dotenv
SQLAlchemy
mysql-connector-python
jwt
pyjwt
